requirements.txt
transformers==4.31.0
python-dotenv==1.0.0
torch==2.0.1
accelerate==0.21.0
xformers==0.0.21